requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic-settings
openai==1.3.0
faiss-cpu==1.7.4
numpy==1.26.2
python-multipart==0.0.6
langchain==0.0.340
tiktoken==0.5.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-telegram-bot==20.7